+++ semgrep/semgrep/evaluation.py
@@ -87,7 +87,9 @@ def _evaluate_single_expression(
                     output_ranges.add(pattern_match.range)
         debug_print(f"after filter `{expression.operator}`: {output_ranges}")
         return output_ranges
-
+    elif expression.operator == OPERATORS.REGEX:
+        # remove all ranges that don't equal the ranges for this pattern
+        return ranges_left.intersection(results_for_pattern)
     else:
         raise NotImplementedError(f"unknown operator {expression.operator}")