+++ semgrep/semgrep/util.py
@@ -1,6 +1,10 @@
+import itertools
 import sys
 from typing import Any
+from typing import Callable
 from typing import Iterable
+from typing import List
+from typing import Tuple
 from urllib.parse import urlparse
 
 global DEBUG
@@ -64,3 +68,9 @@ def set_flags(debug: bool, quiet: bool) -> None:
     if quiet:
         QUIET = True
         debug_print("QUIET is on")
+
+
+def partition(pred: Callable, iterable: Iterable) -> Tuple[List, List]:
+    """E.g. partition(is_odd, range(10)) -> 1 3 5 7 9  and  0 2 4 6 8"""
+    i1, i2 = itertools.tee(iterable)
+    return list(filter(pred, i1)), list(itertools.filterfalse(pred, i2))